+++ jvm.com/classfile/constant-info.go
@@ -16,3 +16,39 @@ const (
 	CONSTANT_MethodType         = 16
 	CONSTANT_InvokeDynamic      = 18
 )
+
+
+func newConstantInfo(tag uint8, cp ConstantPool) ConstantInfo {
+	switch tag {
+	case CONSTANT_Integer:
+		return &ConstantIntegerInfo{}
+	case CONSTANT_Float:
+		return &ConstantFloatInfo{}
+	case CONSTANT_Long:
+		return &ConstantLongInfo{}
+	case CONSTANT_Double:
+		return &ConstantDoubleInfo{}
+	case CONSTANT_Utf8:
+		return &ConstantUtf8Info{}
+	case CONSTANT_String:
+		return &ConstantStringInfo{cp: cp}
+	case CONSTANT_Class:
+		return &ConstantClassInfo{cp: cp}
+	case CONSTANT_Fieldref:
+		return &ConstantFieldrefInfo{ConstantMemberrefInfo{cp: cp}}
+	case CONSTANT_Methodref:
+		return &ConstantMethodrefInfo{ConstantMemberrefInfo{cp: cp}}
+	case CONSTANT_InterfaceMethodref:
+		return &ConstantInterfaceMethodrefInfo{ConstantMemberrefInfo{cp: cp}}
+	case CONSTANT_NameAndType:
+		return &ConstantNameAndTypeInfo{}
+	case CONSTANT_MethodType:
+		return &ConstantMethodTypeInfo{}
+	case CONSTANT_MethodHandle:
+		return &ConstantMethodHandleInfo{}
+	case CONSTANT_InvokeDynamic:
+		return &ConstantInvokeDynamicInfo{}
+	default:
+		panic("java.lang.ClassFormatError: constant pool tag!")
+	}
+}
