+++ jvm.com/classfile/constant-pool.go
@@ -1,5 +1,7 @@
 package classfile
 
+import "fmt"
+
 type ConstantInfo interface {
 	readInfo(read *ClassReader)
 }
@@ -29,7 +31,7 @@ func (self ConstantPool) getConstantInfo(index uint16) ConstantInfo {
 		return cpInfo
 	}
 
-	panic("Invalid constant pool index")
+	panic(fmt.Errorf("Invalid constant pool index: %v!", index))
 }
 
 func (self *ConstantPool) getNameAndType(index uint16) (string, string) {
@@ -58,39 +60,4 @@ func readConstantInfo(reader *ClassReader, cp ConstantPool) ConstantInfo {
 	c.readInfo(reader)
 
 	return c
-}
-
-func newConstantInfo(tag uint8, cp ConstantPool) ConstantInfo {
-	switch tag {
-	case CONSTANT_Integer:
-		return &ConstantIntegerInfo{}
-	case CONSTANT_Float:
-		return &ConstantFloatInfo{}
-	case CONSTANT_Long:
-		return &ConstantLongInfo{}
-	case CONSTANT_Double:
-		return &ConstantDoubleInfo{}
-	case CONSTANT_Utf8:
-		return &ConstantUft8Info{}
-	case CONSTANT_String:
-		return &ConstantStringInfo{cp: cp}
-	case CONSTANT_Class:
-		return &ConstantClassInfo{cp: cp}
-	case CONSTANT_Fieldref:
-		return &ConstantFieldrefInfo{ConstantMemberrefInfo{cp: cp}}
-	case CONSTANT_Methodref:
-		return &ConstantMethodrefInfo{ConstantMemberrefInfo{cp: cp}}
-	case CONSTANT_InterfaceMethodref:
-		return &ConstantInterfaceMethodrefInfo{ConstantMemberrefInfo{cp: cp}}
-	case CONSTANT_NameAndType:
-		return &ConstantNameAndTypeInfo{}
-	case CONSTANT_MethodType:
-		return &ConstantMethodTypeInfo{}
-	case CONSTANT_MethodHandle:
-		return &ConstantMethodHandleInfo{}
-	case CONSTANT_InvokeDynamic:
-		return &ConstantInvokeDynamicInfo{}
-	default:
-		panic("java.lang.ClassFormatError: constant pool tag!")
-	}
 }
