+++ jvm.com/classfile/cp-invoke-dynamic.go
@@ -1,7 +1,25 @@
 package classfile
 
+type ConstantMethodHandleInfo struct {
+	referenceKind  uint8
+	referenceIndex uint16
+}
+
+func (self *ConstantMethodHandleInfo) readInfo(reader *ClassReader) {
+	self.referenceKind = reader.readUint8()
+	self.referenceIndex = reader.readUint16()
+}
+
+type ConstantMethodTypeInfo struct {
+	descriptorIndex uint16
+}
+
+func (self *ConstantMethodTypeInfo) readInfo(reader *ClassReader) {
+	self.descriptorIndex = reader.readUint16()
+}
+
+
 type ConstantInvokeDynamicInfo struct {
-	cp                       ConstantPool
 	bootstrapMethodAttrIndex uint16
 	nameAndTypeIndex         uint16
 }
@@ -9,17 +27,4 @@ type ConstantInvokeDynamicInfo struct {
 func (self *ConstantInvokeDynamicInfo) readInfo(reader *ClassReader) {
 	self.bootstrapMethodAttrIndex = reader.readUint16()
 	self.nameAndTypeIndex = reader.readUint16()
-}
-
-func (self *ConstantInvokeDynamicInfo) NameAndType() (string, string) {
-	return self.cp.getNameAndType(self.nameAndTypeIndex)
-}
-
-func (self *ConstantInvokeDynamicInfo) BootstrapMethodInfo() (uint16, []uint16) {
-	bmAttr := self.cp.cf.BootstrapMethodsAttribute()
-
-	bm := bmAttr.bootstrapMethods[self.bootstrapMethodAttrIndex]
-
-	return bm.bootstrapMethodRef, bm.bootstrapArguments
-
 }
