+++ jvm.com/classpath/entry-composite.go
@@ -1,6 +1,9 @@
 package classpath
 
-import "strings"
+import (
+	"strings"
+	"errors"
+)
 
 type CompositeEntry []Entry
 func newCompositeEntry(pathList string)  CompositeEntry{
@@ -23,6 +26,8 @@ func (self CompositeEntry) readClass(className string) ([]byte, Entry, error){
 			return  data, from, nil
 		}
 	}
+
+	return nil, nil, errors.New("Class not found" + className)
 }
 
 func (self CompositeEntry) String() string {
