+++ jvm.com/classpath/entry.go
@@ -7,7 +7,6 @@ import (
 
 const pathListSeparator = string(os.PathListSeparator)
 
-// 2.3.1
 type Entry interface {
 	readClass (className string) ([]byte, Entry, error)
 
@@ -26,7 +25,7 @@ func newEntry(path string) Entry {
 
 	if strings.HasSuffix(path, ".jar") || strings.HasSuffix(path, ".JAR") ||
 		strings.HasSuffix(path, ".zip") || strings.HasSuffix(path, ".ZIP") {
-			return  newZipEntry()
+			return  newZipEntry(path)
 	}
 
 	return  newDirEntry(path)
